% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fun20210501.R
\name{partial.perm.test}
\alias{partial.perm.test}
\title{Conducting Partial Permutation Test}
\usage{
partial.perm.test(
  Y,
  X,
  Z,
  cov.mat = NULL,
  kernel.list = list(name = "Gaussian", alpha = 0.5, singular = 10^(-5)),
  perm.max = 500,
  threshold = 10^(-5),
  perm.choice = "GP",
  sig.level = 0.05,
  alternative = "GP.pseudo",
  print.opt = perm.max + 1,
  bandwidth.choice = "MLE.all",
  initial.grid = FALSE,
  grid = seq(from = 0.1, to = 5, by = 1),
  print.grid = length(grid)^ncol(X) + 1
)
}
\arguments{
\item{Y}{An \eqn{n} dimensional vector of outcomes}

\item{X}{An \eqn{n} by \eqn{d} matrix of covariates.}

\item{Z}{An \eqn{n} dimensional vector of group indicators, taking values 1, 2, ..., \eqn{H}.}

\item{cov.mat}{An \eqn{n} by \eqn{n} (estimated) error covariance matrix up to a postive scale. It is set to be NULL if it equals identity matrix.}

\item{kernel.list}{A list object for the choice of kernel matrix. For example, it equals list(name = "Gaussian", alpha = 0.5, singular = 10^(-5)) for Gaussian kernel, where alpha is the parameter for bandwidth and singular is used to avoid potential singularity of kernel matrix; it equals list(name = "polynomial", p = 2, singular = 0) for polynomial kernel, where p is for the degree of freedom.}

\item{perm.max}{A numerical object for the number of permutations.}

\item{perm.choice}{A character object specifying how to choose the permutation size. It can take value "GP" and "fix".}

\item{sig.level}{A numerical object for the significance level.}

\item{alternative}{A character object specifying the choice of test statistics. It can take values "GP.equal", "GP.unequal" and "GP.pseudo" for likelihood ratio statistics with different alternative Gaussian process models, and "GP.reg" for comparing mean squared errors from pooled and group-specific kernel regression.}

\item{bandwidth.choice}{A character object specifying how to choose the bandwidth parameter for the kernel function. It can take value "MLE.all" and "fix".}

\item{initial.grid}{A logical object indicating whether to perform grid search for finding the bandwidth parameter that maximizes the marginal likelihood.}

\item{grid}{A vector object indicating the grid to search over when initial.grid is TRUE.}
}
\description{
Get the p-value from the partial permutation test.
}
